Factor out hook for receiving messages from the extension
